Fetch data only once, closes #5
